requirements.txt
flask
python-dotenv
sentence-transformers
faiss-cpu
numpy

